requirements.txt
# Core dependencies
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data handling
datasets>=2.14.0
pandas>=2.0.0
nltk>=3.8.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
tqdm>=4.65.0
rich>=13.0.0

# Experiment tracking
mlflow>=2.7.0

# Evaluation metrics
seaborn>=0.12.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
loguru>=0.7.0
